MCR-1349 #comment change type of getMaxUploadSize() from int to long
git-svn-id: https://server.mycore.de/svn/mycore/trunk@36085 7089d063-bc14-496d-ac90-53de1872eb6c

+++ mycore-sword/src/main/java/org/mycore/sword/MCRSwordConfigurationDefault.java
@@ -53,8 +53,8 @@ public class MCRSwordConfigurationDefault implements SwordConfiguration {
     }
 
     @Override
-    public int getMaxUploadSize() {
-        return MCRConfiguration.instance().getInt("MCR.SWORD.Max.Uploaded.File.Size");
+    public long getMaxUploadSize() {
+        return MCRConfiguration.instance().getLong("MCR.SWORD.Max.Uploaded.File.Size");
     }
 
     @Override
